# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Gin FAST API basic Setup.",
        "title": "Gin API Template",
        "contact": {
            "name": "Chethan NV",
            "url": "https://chetuachar.github.io",
            "email": "chethuchethan72@gamil.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "do health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check the gin server health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "description": "Returns the last 100 log entries in HTML format. Requires a password for authentication.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "View last 100 logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Password for log access",
                        "name": "passwd",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML page with logs",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to load logs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/hi": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "Test api sending",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Auth-Token",
            "in": "header"
        }
    }
}